remove warnings

+++ ZZQCommonUtils/UIViewController+ZZQExtension.m
@@ -27,7 +27,7 @@
         NSInvocation *invocation = [NSInvocation invocationWithMethodSignature:[UIDevice instanceMethodSignatureForSelector:selector]];
         [invocation setSelector:selector];
         [invocation setTarget:[UIDevice currentDevice]];
-        int val = orientation;
+        NSInteger val = orientation;
         // 从2开始是因为0 1 两个参数已经被selector和target占用
         [invocation setArgument:&val atIndex:2];
         [invocation invoke];
